requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
astroid==2.11.2
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
bidict==0.21.4
bleach==5.0.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.4
commonmark==0.9.1
cryptography==36.0.2
cssselect==1.1.0
decorator==5.1.1
dicttoxml==1.7.4
dill==0.3.4
dnspython==2.0.0
docutils==0.18.1
Faker==13.3.5
Flask==2.0.3
Flask-Cors==3.0.10
Flask-SocketIO==5.1.1
Flask-SQLAlchemy==2.5.1
frozenlist==1.3.0
future==0.18.2
geocoder==1.38.1
geoip2==4.5.0
gevent==21.12.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
idna==3.3
importlib-metadata==4.11.3
ip2geotools==0.1.6
IP2Location==8.7.2
ipaddress==1.0.23
isort==5.10.1
itsdangerous==2.1.1
jeepney==0.8.0
Jinja2==3.0.3
keyring==23.5.0
lazy-object-proxy==1.7.1
lxml==4.8.0
MarkupSafe==2.1.1
maxminddb==2.2.0
mccabe==0.7.0
multidict==6.0.2
ndg-httpsclient==0.5.1
outcome==1.1.0
packaging==21.3
pip-review==1.1.1
pkginfo==1.8.2
platformdirs==2.5.1
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.11.2
PyJWT==2.3.0
pylint==2.13.5
PyNaCl==1.5.0
pyOpenSSL==22.0.0
pyparsing==3.0.7
pyquery==1.4.3
pyshorteners==1.0.1
python-dateutil==2.8.2
python-engineio==4.3.1
python-socketio==5.5.2
ratelim==0.1.6
readme-renderer==34.0
requests==2.27.1
requests-toolbelt==0.9.1
rfc3986==2.0.0
rich==12.2.0
SecretStorage==3.3.1
selenium==4.1.3
simple-websocket==0.5.1
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.32
timedelta==2020.12.3
toml==0.10.2
tomli==2.0.1
tqdm==4.64.0
trio==0.20.0
trio-websocket==0.9.2
twine==4.0.0
typed-ast==1.5.2
typing==3.7.4.3
typing-extensions==4.1.1
urllib3==1.26.9
uuid==1.30
webencodings==0.5.1
Werkzeug==2.0.3
wrapt==1.14.0
wsproto==1.1.0
yarl==1.7.2
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0
